hs_core_server/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.10.0
attrs==25.4.0
babel==2.17.0
black==25.9.0
certifi==2025.10.5
cffi==2.0.0
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.4
click==8.3.0
distro==1.9.0
Django==5.2.7
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
docxcompose==1.4.0
dotenv==0.9.9
drf-nested-routers==0.95.0
drf-spectacular==0.28.0
drf-yasg==1.21.11
et_xmlfile==2.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
inflection==0.5.1
isort==7.0.0
jiter==0.11.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lxml==6.0.2
minio==7.2.18
msgpack==1.1.2
mypy_extensions==1.1.0
openai==2.5.0
openpyxl==3.1.5
packaging==25.0
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.0
psycopg2-binary==2.9.11
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.3
pydantic_core==2.41.4
PyJWT==2.10.1
pypdf==6.1.1
python-docx==1.2.0
python-dotenv==1.1.1
python-json-logger==4.0.0
pytokens==0.2.0
pytz==2025.2
PyYAML==6.0.3
redis==6.4.0
referencing==0.37.0
reportlab==4.4.4
requests==2.32.5
rpds-py==0.27.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
websockets==15.0.1
